Ho risolto il bug della rimozione di album e fotografi in uso. Ho aggiunto ovunque il bottone per sloggare.

+++ src/main/java/com/piniscarlatti/siw/controller/FunzionarioAlbumController.java
@@ -42,10 +42,12 @@ public class FunzionarioAlbumController implements WebMvcConfigurer {
 
     //photographers/{id}/album/{idAlbum}/delete
     @GetMapping("/{idAlbum}/delete")
-    public String deletePhotographers(@PathVariable("id") Long id, @PathVariable("idAlbum") Long idAlbum, Model model) {
+    public String deleteAlbum(@PathVariable("id") Long id, @PathVariable("idAlbum") Long idAlbum, Model model) {
         Fotografo fotografo = fotografoService.getFotografoById(id);
         Album album = albumService.getAlbumById(idAlbum);
-        albumService.delete(album);
+        if(!fotografoService.ciSonoFotoAlbumInOrdini(id,idAlbum) && !fotografoService.ciSonoFotoAlbumInCarrelli(id,idAlbum)) {
+            albumService.delete(album);
+        }
         model.addAttribute("albums", albumService.getAlbumsByFotografo(fotografo));
         return "redirect:/photographers/{id}/album";
     }
